backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
cryptography==3.4.7
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.7
django-classy-tags==2.0.0
django-cms==3.9.0
django-colorfield==0.4.5
django-cors-headers==3.3.0
django-formtools==2.3
django-heroku==0.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
Django-Verify-Email==1.0.7
djangocms-admin-style==2.0.2
djangocms-snippet==3.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
gunicorn==20.1.0
jwt==1.2.0
package-name==0.1
Pillow==8.4.0
platformdirs==2.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2020.1
six==1.16.0
sqlparse==0.3.1
toml==0.10.2
virtualenv==20.7.2
whitenoise==5.3.0
